Eliminate pyright errors (#2062)
* Add Self annotation to gui.rect.move
* Skip dot reference in gui.Rect.moveby unpacking
* Add return annotations on gui.Rect.properties
* Annotate gui.Rect.resize argument types
* Add gui.Rect.resize docstring annotations
* Document gui.Rect.size
* Revert mention of pixels in gui.Rect.resize
* Add docstrings for edge properties with mention of axis alignment
* Use float values to make type checkers happy
* Add arcade.types.AsFloat alias
* Use AsFloat in gui.Rect.move signature
* Annotate gui.Rect.collide_with_point
* Add docstring for gui.Rect.collide_with_point
* Expand docstring for gui.Rect.scale
* Make gui.Rect.scale's rounding adjustable
* use AsFloat in gui.Rect.align_right
* Type hint gui.Rect.align_center
* Use AsFloat in align_center_x and align_center_y
* Type hint gui.Rect.min_size
* Type hint gui.Rect.max_size
* Skip dot access in gui.Rect.max_size
* Add return annotation on gui.Rect.union
* Add tests for gui.rect.Scale
* Leave off Self magic for now
* Fix Callable first argument for now
* Make UIDropDown's do_layout None-aware
* Remove unused import
* Make LAYOUT_OFFSET a float to implicit type convert to make pyright happy
* Fix copy and paste issue in rect unit tests
* Add float conversion to make pyright happy
* Convert to int when interfacing with IncrementalTextLayout
* Round to int locally on UITextArea
* Return floats from arcade.Text lrbt props

+++ arcade/text.py
@@ -442,28 +442,28 @@ class Text:
         return self._label.content_height
 
     @property
-    def left(self) -> int:
+    def left(self) -> float:
         """
         Pixel location of the left content border.
         """
         return self._label.left
 
     @property
-    def right(self) -> int:
+    def right(self) -> float:
         """
         Pixel location of the right content border.
         """
         return self._label.right
 
     @property
-    def top(self) -> int:
+    def top(self) -> float:
         """
         Pixel location of the top content border.
         """
         return self._label.top
 
     @property
-    def bottom(self) -> int:
+    def bottom(self) -> float:
         """
         Pixel location of the bottom content border.
         """
